Consolidate reasoning content tests into test_enable_thinking.py
Move TestReasoningContentAPI tests (separate_reasoning feature) from
test_reasoning_content.py into test_enable_thinking.py, which already
launches Qwen3-30B-A3B with --reasoning-parser qwen3.

Delete test_reasoning_content.py entirely:
- TestReasoningContentAPI: merged into test_enable_thinking.py
- TestReasoningContentWithoutParser: low-value edge case (no-parser +
  separate_reasoning=True), can be a CPU unit test later if needed

This eliminates 2 dedicated server launches (DeepSeek-R1 with and
without reasoning parser).
